impl new BlockDevice for Card

+++ src/fat32.rs
@@ -5,21 +5,21 @@ use crate::sdcard::{CmdError, Card};
 impl BlockDevice for Card {
     type Error = CmdError;
 
-    fn read(&self, buf: &mut [u8], address: u32, number_of_blocks: u32) -> Result<(), Self::Error> {
+    fn read(&self, buf: &mut [u8], address: usize, number_of_blocks: usize) -> Result<(), Self::Error> {
         if number_of_blocks == 1 {
-            self.read_block(buf, address)?
+            self.read_block(buf, address as u32)?
         } else {
-            self.read_multi_blocks(buf, address, number_of_blocks)?
+            self.read_multi_blocks(buf, address as u32, number_of_blocks as u32)?
         }
 
         Ok(())
     }
 
-    fn write(&self, buf: &[u8], address: u32, number_of_blocks: u32) -> Result<(), Self::Error> {
+    fn write(&self, buf: &[u8], address: usize, number_of_blocks: usize) -> Result<(), Self::Error> {
         if number_of_blocks == 1 {
-            self.write_block(buf, address)?
+            self.write_block(buf, address as u32)?
         } else {
-            self.write_multi_blocks(buf, address, number_of_blocks)?
+            self.write_multi_blocks(buf, address as u32, number_of_blocks as u32)?
         }
 
         Ok(())
